requirements.txt
anchorpy==0.20.1
anchorpy-core==0.2.0
anyio==4.6.0
attrs==24.2.0
base58==2.1.1
based58==0.1.1
borsh-construct==0.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
construct==2.10.68
construct-typing==0.5.6
cryptography==43.0.1
exceptiongroup==1.2.2
frozendict==2.3.10
future==1.0.0
grpcio==1.65.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
invoke==2.2.0
jito_py_rpc==0.1.0
jsonalias==0.1.1
jsonrpcclient==4.0.3
jupiter-python-sdk==0.0.2.0
moralis==0.1.48
more-itertools==8.14.0
numpy==2.1.2
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
psycopg==3.1.20
psycopg-pool==3.2.0
psycopg2==2.9.10
pycparser==2.22
pyheck==0.1.5
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==12.0.0
pytz==2024.2
qrcode==7.4.2
requests==2.32.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
solana==0.34.0
solders==0.21.0
sumtypes==0.1a6
tabulate==0.9.0
toml==0.10.2
toolz==0.11.2
tornado==6.4.1
typing_extensions==4.3.0
tzdata==2024.2
urllib3==1.26.20
websockets==10.4
zenrows==1.3.2
